refactor(data-model): clean data model
