+++ cockpit_api/src/main/java/com/cockpit/api/model/dto/jira/AvatarUrls.java
@@ -1,27 +1,21 @@
 package com.cockpit.api.model.dto.jira;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.fasterxml.jackson.annotation.*;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
         "48x48",
-        "24x24",
-        "16x16",
-        "32x32",
 })
 public class AvatarUrls {
 
     @JsonProperty("48x48")
     private String bigAvatar;
-    @JsonProperty("24x24")
-    private String smallAvatar;
-    @JsonProperty("16x16")
-    private String verySmallAvatar;
-    @JsonProperty("32x32")
-    private String mediumAvatar;
 
+    @JsonIgnore
+    private Map<String, Object> additionalProperties = new HashMap<>();
     @JsonProperty("48x48")
     public String getBigAvatar() {
         return bigAvatar;
@@ -32,33 +26,13 @@ public class AvatarUrls {
         this.bigAvatar = bigAvatar;
     }
 
-    @JsonProperty("24x24")
-    public String getSmallAvatar() {
-        return smallAvatar;
+    @JsonAnyGetter
+    public Map<String, Object> getAdditionalProperties() {
+        return this.additionalProperties;
     }
 
-    @JsonProperty("24x24")
-    public void setSmallAvatar(String smallAvatar) {
-        this.smallAvatar = smallAvatar;
-    }
-
-    @JsonProperty("16x16")
-    public String getVerySmallAvatar() {
-        return verySmallAvatar;
-    }
-
-    @JsonProperty("16x16")
-    public void setVerySmallAvatar(String verySmallAvatar) {
-        this.verySmallAvatar = verySmallAvatar;
-    }
-
-    @JsonProperty("32x32")
-    public String getMediumAvatar() {
-        return mediumAvatar;
-    }
-
-    @JsonProperty("32x32")
-    public void setMediumAvatar(String mediumAvatar) {
-        this.mediumAvatar = mediumAvatar;
+    @JsonAnySetter
+    public void setAdditionalProperty(String name, Object value) {
+        this.additionalProperties.put(name, value);
     }
 }
