--- cockpit_api/src/main/java/com/cockpit/api/model/dto/jira/Changelog.java
@@ -1,66 +1,0 @@
-package com.cockpit.api.model.dto.jira;
-
-import com.fasterxml.jackson.annotation.*;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "startAt",
-        "maxResults",
-        "total",
-})
-public class Changelog {
-
-    @JsonProperty("startAt")
-    private Integer startAt;
-    @JsonProperty("maxResults")
-    private Integer maxResults;
-    @JsonProperty("total")
-    private Integer total;
-    @JsonIgnore
-    private Map<String, Object> additionalProperties = new HashMap<>();
-
-    @JsonProperty("startAt")
-    public Integer getStartAt() {
-        return startAt;
-    }
-
-    @JsonProperty("startAt")
-    public void setStartAt(Integer startAt) {
-        this.startAt = startAt;
-    }
-
-    @JsonProperty("maxResults")
-    public Integer getMaxResults() {
-        return maxResults;
-    }
-
-    @JsonProperty("maxResults")
-    public void setMaxResults(Integer maxResults) {
-        this.maxResults = maxResults;
-    }
-
-    @JsonProperty("total")
-    public Integer getTotal() {
-        return total;
-    }
-
-    @JsonProperty("total")
-    public void setTotal(Integer total) {
-        this.total = total;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-}
-
