+++ cockpit_api/src/main/java/com/cockpit/api/model/dto/jira/Issue.java
@@ -6,39 +6,21 @@ import java.util.Map;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "expand",
         "id",
-        "self",
         "key",
-        "changelog",
         "fields"
 })
 public class Issue {
 
-    @JsonProperty("expand")
-    private String expand;
     @JsonProperty("id")
     private String id;
-    @JsonProperty("self")
-    private String self;
+
     @JsonProperty("key")
     private String key;
-    @JsonProperty("changelog")
-    private Changelog changelog;
     @JsonProperty("fields")
     private Fields fields;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<>();
-
-    @JsonProperty("expand")
-    public String getExpand() {
-        return expand;
-    }
-
-    @JsonProperty("expand")
-    public void setExpand(String expand) {
-        this.expand = expand;
-    }
 
     @JsonProperty("id")
     public String getId() {
@@ -50,34 +32,13 @@ public class Issue {
         this.id = id;
     }
 
-    @JsonProperty("self")
-    public String getSelf() {
-        return self;
-    }
-
-    @JsonProperty("self")
-    public void setSelf(String self) {
-        this.self = self;
-    }
-
     @JsonProperty("key")
     public String getKey() {
         return key;
     }
-
     @JsonProperty("key")
     public void setKey(String key) {
         this.key = key;
-    }
-
-    @JsonProperty("changelog")
-    public Changelog getChangelog() {
-        return changelog;
-    }
-
-    @JsonProperty("changelog")
-    public void setChangelog(Changelog changelog) {
-        this.changelog = changelog;
     }
 
     @JsonProperty("fields")
