+++ cockpit_api/src/main/java/com/cockpit/api/model/dto/jira/IssueType.java
@@ -7,39 +7,20 @@ import java.util.Map;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "self",
         "id",
         "description",
-        "iconUrl",
         "name",
-        "subtask"
 })
 public class IssueType {
 
-    @JsonProperty("self")
-    private String self;
     @JsonProperty("id")
     private String id;
     @JsonProperty("description")
     private String description;
-    @JsonProperty("iconUrl")
-    private String iconUrl;
     @JsonProperty("name")
     private String name;
-    @JsonProperty("subtask")
-    private Boolean subtask;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<>();
-
-    @JsonProperty("self")
-    public String getSelf() {
-        return self;
-    }
-
-    @JsonProperty("self")
-    public void setSelf(String self) {
-        this.self = self;
-    }
 
     @JsonProperty("id")
     public String getId() {
@@ -61,16 +42,6 @@ public class IssueType {
         this.description = description;
     }
 
-    @JsonProperty("iconUrl")
-    public String getIconUrl() {
-        return iconUrl;
-    }
-
-    @JsonProperty("iconUrl")
-    public void setIconUrl(String iconUrl) {
-        this.iconUrl = iconUrl;
-    }
-
     @JsonProperty("name")
     public String getName() {
         return name;
@@ -79,16 +50,6 @@ public class IssueType {
     @JsonProperty("name")
     public void setName(String name) {
         this.name = name;
-    }
-
-    @JsonProperty("subtask")
-    public Boolean getSubtask() {
-        return subtask;
-    }
-
-    @JsonProperty("subtask")
-    public void setSubtask(Boolean subtask) {
-        this.subtask = subtask;
     }
 
     @JsonAnyGetter
