+++ cockpit_api/src/main/java/com/cockpit/api/model/dto/jira/Project.java
@@ -8,52 +8,20 @@ import java.util.Map;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "expand",
-    "self",
     "id",
     "key",
     "name",
-    "projectTypeKey",
-    "simplified"
 })
 public class Project {
 
-    @JsonProperty("expand")
-    private String expand;
-    @JsonProperty("self")
-    private String self;
     @JsonProperty("id")
     private String id;
     @JsonProperty("key")
     private String key;
     @JsonProperty("name")
     private String name;
-    @JsonProperty("projectTypeKey")
-    private String projectTypeKey;
-    @JsonProperty("simplified")
-    private Boolean simplified;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<>();
-
-    @JsonProperty("expand")
-    public String getExpand() {
-        return expand;
-    }
-
-    @JsonProperty("expand")
-    public void setExpand(String expand) {
-        this.expand = expand;
-    }
-
-    @JsonProperty("self")
-    public String getSelf() {
-        return self;
-    }
-
-    @JsonProperty("self")
-    public void setSelf(String self) {
-        this.self = self;
-    }
 
     @JsonProperty("id")
     public String getId() {
@@ -83,28 +51,6 @@ public class Project {
     @JsonProperty("name")
     public void setName(String name) {
         this.name = name;
-    }
-
-  
-
-    @JsonProperty("projectTypeKey")
-    public String getProjectTypeKey() {
-        return projectTypeKey;
-    }
-
-    @JsonProperty("projectTypeKey")
-    public void setProjectTypeKey(String projectTypeKey) {
-        this.projectTypeKey = projectTypeKey;
-    }
-
-    @JsonProperty("simplified")
-    public Boolean getSimplified() {
-        return simplified;
-    }
-
-    @JsonProperty("simplified")
-    public void setSimplified(Boolean simplified) {
-        this.simplified = simplified;
     }
 
     @JsonAnyGetter
