+++ cockpit_api/src/main/java/com/cockpit/api/model/dto/jira/Status.java
@@ -7,37 +7,21 @@ import java.util.Map;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "self",
         "description",
-        "iconUrl",
         "name",
         "id",
         "statusCategory"
 })
 public class Status {
 
-    @JsonProperty("self")
-    private String self;
     @JsonProperty("description")
     private String description;
-    @JsonProperty("iconUrl")
-    private String iconUrl;
     @JsonProperty("name")
     private String name;
     @JsonProperty("id")
     private String id;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<>();
-
-    @JsonProperty("self")
-    public String getSelf() {
-        return self;
-    }
-
-    @JsonProperty("self")
-    public void setSelf(String self) {
-        this.self = self;
-    }
 
     @JsonProperty("description")
     public String getDescription() {
@@ -47,16 +31,6 @@ public class Status {
     @JsonProperty("description")
     public void setDescription(String description) {
         this.description = description;
-    }
-
-    @JsonProperty("iconUrl")
-    public String getIconUrl() {
-        return iconUrl;
-    }
-
-    @JsonProperty("iconUrl")
-    public void setIconUrl(String iconUrl) {
-        this.iconUrl = iconUrl;
     }
 
     @JsonProperty("name")
